자기 아이디 게시물 삭제 및 자기 아이디 뜨게 하기

+++ backend/src/main/java/com/example/backend/controller/borad/BoardController.java
@@ -4,6 +4,7 @@ import com.example.backend.dto.borad.Board;
 import com.example.backend.service.borad.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
@@ -36,15 +37,23 @@ public class BoardController {
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity<Map<String, Object>> delete(@PathVariable int id) {
-        if (service.remove(id)) {
-            return ResponseEntity.ok()
-                    .body(Map.of("message", Map.of("type", "success"
-                            , "text", STR."\{id}번 게시글이 삭제되었습니다.")));
-        } else {
-            return ResponseEntity.internalServerError()
-                    .body(Map.of("message", Map.of("type", "error"
-                            , "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> delete(@PathVariable int id,
+                                                      Authentication authentication) {
+        if (service.hashCode(id,authentication)){
+            if (service.remove(id)) {
+                return ResponseEntity.ok()
+                        .body(Map.of("message", Map.of("type", "success"
+                                , "text", STR."\{id}번 게시글이 삭제되었습니다.")));
+            } else {
+                return ResponseEntity.internalServerError()
+                        .body(Map.of("message", Map.of("type", "error"
+                                , "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+            }
+        }else {
+            return ResponseEntity.status(403)
+                    .body(Map.of("message", Map.of("type", "error",
+                            "text","권한 없음")));
         }
     }
 
@@ -72,6 +81,7 @@ public class BoardController {
     }
 
     @PostMapping("add")
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<Map<String, Object>> add(@RequestBody Board board,
                                                    Authentication authentication) {
